Add resources and JSP

+++ src/main/java/dal/AuthorDao.java
@@ -1,14 +1,13 @@
 package dal;
 
 import dao.Author;
-import dao.Book;
 
 import java.util.List;
 
 public interface AuthorDao {
     void save(Author entity);
 
-    Author findById(Long authorId);
+    Author findById(int authorId);
 
     Author findByFirstNameAndLastName(String firstName, String lastName);
